test(sidebar): recover the orphaned spending-origin pin
